+++ packages/noir_signals/test/hooks/replacement_lifecycle_test.dart
@@ -1,8 +1,8 @@
-import 'package:noir/hooks.dart';
 import 'package:noir/noir.dart';
+import 'package:noir_signals/noir_signals.dart';
 import 'package:test/test.dart';
 
-import '../helpers/test_element_host.dart';
+import '../helpers/noir_test_helpers.dart';
 
 void main() {
   test('keyed owners rewire later listeners before old disposal', () {